Fix formatting
For comments, add a space between the '#' and the start of the comment.

For function arguments, remove extra whitespace around '='.

For dictionary add extra space between colon and value.

Wrap long comment lines.

Move if statement body on the next on the next line instead of the same
line as the if statement.

Add extra space after comma in line max() arguments.

Make XboxController() last argument on new line so not as long.